fix: tighten editor shortcut modifier mapping
